+++ src/demo/sections/Inputs.vue
@@ -4,15 +4,21 @@
     <div v-for="state in states" :key="state" class="my-2 setWidth">
       <Input
         v-model="state.val"
+        id="123"
+        type="state.type"
+        name="input"
+        autocomplete="on"
+        class="ml-3"
+        :placeholder="state.placeholder"
+        required
         :disabled="state.state == 'disabled'"
         :label="state.label"
         @invalid="handleInvalid"
         :subtitle="state.subtitle"
-        required
-        :validityIcon="false"
+        :validityIcon="true"
       >
         <template v-slot:prepend>
-          <i class="pi pi-Search pi-2x"></i>
+          <i class="pi-Search pi-2x"></i>
         </template>
         <template v-slot:append>
           <Toggle :disabled="state.state == 'disabled'" />
@@ -32,8 +38,9 @@ export default class Inputs extends Vue {
   minVal = 0 as number
   maxVal = 10 as number
   states = [
-    { state: 'default', val: 'Hello', subtitle: 'Default subtitle' },
-    { state: 'disabled', val: 'Hello again', label: 'Disabled' }
+    { state: 'default', val: 'hello', subtitle: 'Default subtitle', label: "test label" },
+    { state: 'placeholder', val: '', subtitle: 'placeholder subtitle', placeholder:"Placeholder" },
+    { state: 'disabled', val: 'Hello again', label: 'Disabled',  }
   ]
 
   mounted(): void {